move some parts of malloc-trace code
