+++ malloc-trace/precomp_inc/precomp.hh
@@ -1,0 +1,33 @@
+/*
+** Copyright 2022 Centreon
+**
+** This file is part of Centreon Engine.
+**
+** Centreon Engine is free software: you can redistribute it and/or
+** modify it under the terms of the GNU General Public License version 2
+** as published by the Free Software Foundation.
+**
+** Centreon Engine is distributed in the hope that it will be useful,
+** but WITHOUT ANY WARRANTY; without even the implied warranty of
+** MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+** General Public License for more details.
+**
+** You should have received a copy of the GNU General Public License
+** along with Centreon Engine. If not, see
+** <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef CMT_PRECOMP_HH
+#define CMT_PRECOMP_HH
+
+#include <sys/syscall.h>
+#include <unistd.h>
+
+#include <map>
+#include <memory>
+#include <mutex>
+
+#include <absl/strings/numbers.h>
+#include <boost/intrusive/set.hpp>
+
+#endif
